Results Logic
condensed booksearch into 1 method thats called with a get and a post in routes.
Added a blank book image for books that dont have an image attribute.
created a book view for viewing the singular instance of a book, that is going to display more information about it.
Created a footer component